# src: equally many limb-optimize APyFixedArray::operator+/-()