### Benchmark:

```Python
from apytypes import *
from time import time
import numpy as np


a = APyFixedArray.from_float(np.random.rand(200, 200) - 0.5, 200, 1)
b = APyFixedArray.from_float(np.random.rand(200, 200) - 0.5, 200, 1)

t1 = time()
for i in range(10000):
    a + b
t2 = time()
print(f'[200 x 200] 200-bit addition: {(t2 - t1)*1e6 / 10000} us per addition')

t1 = time()
for i in range(10000):
    a - b 
t2 = time()
print(f'[200 x 200] 200-bit subtraction: {(t2 - t1)*1e6 / 10000} us per subtraction')
```

### Before (b01c1ad):

```
[200 x 200] 200-bit addition: 1035.451579093933 us per addition
[200 x 200] 200-bit subtraction: 986.5080833435059 us per subtraction
```

### After:

```
[200 x 200] 200-bit addition: 148.31440448760986 us per addition
[200 x 200] 200-bit subtraction: 162.70325183868408 us per subtraction
```